+++ app/Filament/Resources/CustomerResource/Pages/Modules.php
@@ -39,6 +39,7 @@ class Modules extends Page implements HasTable
 
     public function table(Table $table): Table
     {
-        return $this->customerScopedResourceTable($table);
+        return $this->customerScopedResourceTable($table)
+            ->headerActions([$this->customerScopedCreateAction('Add Module')]);
     }
 }