+++ app/Filament/Resources/CustomerResource/Pages/Subscription.php
@@ -40,6 +40,7 @@ class Subscription extends Page implements HasTable
 
     public function table(Table $table): Table
     {
-        return $this->customerScopedResourceTable($table);
+        return $this->customerScopedResourceTable($table)
+            ->headerActions([$this->customerScopedCreateAction('Add Subscription')]);
     }
 }